# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Ambulanzsystem V1 API
  version: 0.1.0
  description: |
    Frozen Stage-0 contract for the Ambulanzsystem V1 (see docs/plan/master-plan.md).
    Locked decisions D1-D8 apply. JSON property names are serialized exactly as
    written here (no automatic camelCase transformation server-side).

    Auth token types (JWT `type` claim): `admin`, `leitstelle`, `user`, `qr`.
    Policies referenced below:
    - AdminOnly: type=admin
    - LeitstelleOrAdmin: type in (admin, leitstelle)
    - AuthenticatedUser: type in (admin, leitstelle, user)
    - TriageWrite: type in (admin, leitstelle, user, qr)

    Canonical triage values: rot | gelb | gruen | schwarz (ASCII, D5).
    `blau` and Unicode `grün` are rejected at the write boundary.
servers:
  - url: http://localhost:5000
    description: local backend
  - url: http://localhost:4010
    description: Prism mock server
security:
  - bearerAuth: []
...
paths:
...
  /api/qr-login:
    post:
      tags: [auth]
      summary: Responder QR login (online-only). Short-lived token, no refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qr_code]
              properties:
                qr_code: { type: string, maxLength: 128 }
      responses:
        "200":
          description: Logged in. Token has type=qr, scoped to the event scene subtree.
          content:
            application/json:
              schema:
                type: object
                required: [status, token]
                properties:
                  status: { type: string, const: ok }
                  token: { type: string }
                  eventSceneId: { type: integer, description: Top-level scene this QR grants access to. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/user-login:
    post:
      tags: [auth]
      summary: Responder username/password login (token type=user).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialsRequest" }
      responses:
        "200":
          description: Logged in.
          content:
            application/json:
              schema:
                type: object
                required: [status, token, refreshToken]
                properties:
                  status: { type: string, const: ok }
                  token: { type: string }
                  refreshToken: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/admin-login:
    post:
      tags: [auth]
      summary: Admin or Leitstelle login. Token type reflects the user's role.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialsRequest" }
      responses:
        "200":
          description: Logged in.
          content:
            application/json:
              schema:
                type: object
                required: [status, token, refreshToken, requiresPasswordChange, role]
                properties:
                  status: { type: string, const: ok }
                  token: { type: string }
                  refreshToken: { type: string }
                  requiresPasswordChange: { type: boolean }
                  role: { $ref: "#/components/schemas/Role" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/refresh-token:
    post:
      tags: [auth]
      summary: Rotate a refresh token (old is revoked, new pair issued).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New token pair.
          content:
            application/json:
              schema:
                type: object
                required: [token, refreshToken]
                properties:
                  token: { type: string }
                  refreshToken: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/logout:
    post:
      tags: [auth]
      summary: Revoke the given refresh token. Policy AuthenticatedUser.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/validate-token:
    post:
      tags: [auth]
      summary: Route-guard check. Policy TriageWrite (any authenticated session).
      responses:
        "200":
          description: Token valid.
          content:
            application/json:
              schema:
                type: object
                required: [isValid]
                properties:
                  isValid: { type: boolean, const: true }
                  role: { type: string, enum: [admin, leitstelle, user, qr] }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/dev-login:
    post:
      tags: [auth]
      summary: DEV ONLY. Password-free login for seeded accounts. 404 unless Development env AND Features__EnableDevLogin.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [admin, user] }
      responses:
        "200":
          description: Logged in (real token, dev environment only).
          content:
            application/json:
              schema:
                type: object
                required: [status, token]
                properties:
                  status: { type: string, const: ok }
                  token: { type: string }
                  refreshToken: { type: string }
        "404": { description: Not available (any non-dev configuration). }

  # ---------------------------------------------------------------- users
...
  /api/login-qr-codes/generate:
    post:
      tags: [login-qr-codes]
      summary: Bulk-generate responder login QR codes for an event. Policy LeitstelleOrAdmin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number, eventSceneId]
              properties:
                number: { type: integer, minimum: 1, maximum: 500 }
                eventSceneId: { type: integer, description: Top-level scene (event) the codes grant access to. }
                expiresInHours: { type: number, default: 12, description: Access window from first login (FR-AUTH-08). }
      responses:
        "201":
          description: Generated codes.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LoginQrCode" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/login-qr-codes:
    get:
      tags: [login-qr-codes]
      summary: List login QR codes. Policy LeitstelleOrAdmin.
      parameters:
        - name: eventSceneId
          in: query
          schema: { type: integer }
        - name: unusedOnly
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Codes.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LoginQrCode" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/login-qr-codes/{id}/revoke:
    post:
      tags: [login-qr-codes]
      summary: Revoke a login QR code early. Policy LeitstelleOrAdmin.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------------------------- patient QR codes
...
  /api/body-parts:
    get:
      tags: [body-parts]
      summary: Full region-toggle body state for a patient (all canonical keys always present). Policy TriageWrite.
      parameters:
        - name: idpatient
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Region-state map, one entry per key in contract/body-regions.json.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BodyParts" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [body-parts]
      summary: Toggle one region's marker. bodyPartId validated against the canonical key list. Publishes realtime update. Policy TriageWrite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idpatient, bodyPartId, isClicked]
              properties:
                idpatient: { type: integer }
                bodyPartId: { type: string, description: One key from contract/body-regions.json. }
                isClicked: { type: boolean }
      responses:
        "200":
          description: Updated full body state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BodyParts" }
        "400": { description: Unknown bodyPartId. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------- audit
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Unauthorized:
      description: Missing/invalid/expired/revoked credentials. Expired access blocks new actions and reads (FR-AUTH-09).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but role not permitted (e.g. responder on admin endpoint, NFR-SEC-02).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Entity not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Login rate limit exceeded (fixed window 10/60s per IP, NFR-SEC-05).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [status, message]
      properties:
        status: { type: string, const: error }
        message: { type: string }
    Role:
      type: string
      enum: [admin, leitstelle, responder]
...
    CredentialsRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
...
    LoginQrCode:
      type: object
      required: [id, qrToken, eventSceneId, createdAt]
      properties:
        id: { type: integer }
        qrToken: { type: string }
        eventSceneId: { type: integer }
        firstLogin: { type: [string, "null"], format: date-time }
        expiresAt: { type: [string, "null"], format: date-time, description: Null until first login; then firstLogin + configured window. }
        expiresInHours: { type: number }
        revokedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
...
    BodyParts:
      type: object
      required: [idpatient, bodyParts]
      properties:
        idpatient: { type: integer }
        bodyParts:
          type: object
          description: One entry per key in contract/body-regions.json; 0 = unmarked, non-zero = marked.
          additionalProperties: { type: integer }
        updatedAt: { type: string, format: date-time }
...